requirements.txt
djangorestframework==3.14.0
psycopg2==2.9.7
drf-yasg==1.21.7
django-cors-headers==4.2.0